requirements.txt
colorlog >= 6.0.0
argon2 >= 23.1.0
pycryptodome >= 3.21.0